List overhaul
* move all the "algorithmic" routines in `mathics.core.lists` to an independent module in `mathics.algorithm.parts`
* cithonize `mathics.algorithm.parts`
* fix all the warnings produced by flake8
* Move `Exception` classes from `mathics.builtin.base` to `mathics.builtin.exceptions` to avoid circular references.
