+++ mathics/builtin/tensors.py
@@ -25,7 +25,7 @@ from mathics.core.systemsymbols import (
 )
 from mathics.core.rules import Pattern
 
-from mathics.builtin.lists import get_part
+from mathics.algorithm.parts import get_part
 
 
 def get_default_distance(p):
